requirements.txt
Pillow==10.4.0
requests>=2.31.0
geopy>=2.4.1